% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_boot.R
\name{div_boot}
\alias{div_boot}
\title{Function to summarize the bootstrapped diversity indices.}
\usage{
div_boot(dfl, index = "shannon", progress = TRUE, ...)
}
\arguments{
\item{dfl}{A list of pixel_boot matrices obtained from
\code{\link[soildiv]{resamp_wrapper}}.}

\item{index}{A value of "simpson", "shannon", "richness" or "count"
(i.e. pixel count).}

\item{...}{Additional arguments to \code{\link[vegan]{diversity}} (only
relevant if index = "simpson" or "shannon")}
}
\description{
The function is a wrapper to \code{\link[vegan]{diversity}}
  that is specifically adapted to handle the output of
  \code{\link[soildiv]{resamp_wrapper}}.
}
