% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_wrapper.R
\name{div_wrapper}
\alias{div_wrapper}
\title{Function to calculate diversity indices.}
\usage{
div_wrapper(df, mask, id_var, index = "shannon", ...)
}
\arguments{
\item{df}{The input dataframe obtained from \code{\link[soildiv]{new_class}}.}

\item{mask}{An integer vector of pixel types that should be ignored in the
diversity estimates.}

\item{id_var}{The name of the lone ID varable (the rest should be
pixel counts), should be in quotations.}

\item{index}{A value of "simpson", "shannon", "richness" or "count"
(i.e. pixel count).}

\item{...}{Additional arguments to \code{\link[vegan]{diversity}} (only
relevant if index = "simpson" or "shannon")}
}
\description{
The function is a wrapper to \code{\link[vegan]{diversity}}
  that is specifically adapted to handle the output of
  \code{\link[soildiv]{new_class}}.
}
