% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get_seed}
\alias{get_seed}
\title{Random seed generator}
\usage{
get_seed()
}
\description{
#' Cool way to set a random seed in a c++ function using R as the base.
Obtained from:
https://www.r-bloggers.com/2018/09/using-rs-set-seed-to-set-seeds-for-use-in-c-c-including-rcpp/
}
