% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_class.R
\name{new_class}
\alias{new_class}
\title{Function to reclassify pixel counts given a pixel matching file.}
\usage{
new_class(df, rcl, GEOID, Crop, Count, New_Crop)
}
\arguments{
\item{df}{A dataframe of pixel counts with pixel counts provided in the
columns with a lone id variable. Similar to:
https://www.nass.usda.gov/Research_and_Science/Cropland/docs/County_Pixel_Count.zip}

\item{rcl}{A dataframe of pixel reclassications. The original crop
classification column. If NULL, the function simply renames the
column names to eliminate any leading zeros in the pixel numbers
(which is critical for the div_mask function).}

\item{GEOID}{The (unquoted) name of the ID column in df.}

\item{Crop}{The (unquoted) name for the original crop types in rcl.}

\item{Count}{The (unquoted) desired variable name describing the pixel counts.}

\item{New_Crop}{The (unquoted) variable name for new crop classifications as
specified in rcl.}
}
\value{
A dataframe of new pixel counts given the provided rcl dataframe.
}
\description{
Function to reclassify pixel counts given a pixel matching file.
}
