% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reclass}
\alias{reclass}
\title{Pixel Reclassification Attempt 1.}
\format{
A data frame with 134 rows and 4 columns.

  \describe{
  \item{Crop}{The original integer crop identification numbers.}
  \item{Description}{The description associated with the identification
    number.}
  \item{New_Crop}{The proposed reclassification.}
  \item{New_Description}{The new description of the reclassified crop.}
  }
}
\usage{
reclass
}
\description{
These pixel reclassifications treat double cropping systems as their own
  crop type.
}
\keyword{datasets}
