% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reclass2}
\alias{reclass2}
\title{Pixel Reclassification Attempt 2.}
\format{
A data frame with 151 rows and 4 columns (extra rows for
  double counts).

  \describe{
  \item{Crop}{The original integer crop identification numbers.}
  \item{Description}{The description associated with the identification
    number.}
  \item{New_Crop}{The proposed reclassification.}
  \item{New_Description}{The new description of the reclassified crop.}
  }
}
\usage{
reclass2
}
\description{
These pixel reclassifications double count pixels associated with double
  cropping systems: one count in each of the individual cropping systems.
}
\keyword{datasets}
