% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_boot.R
\name{resamp}
\alias{resamp}
\title{Bootstrap simulation function.}
\usage{
resamp(x, n)
}
\arguments{
\item{x}{An integer vector of length q representing crop pixel counts.}

\item{n}{The desired number of boostrap simulations.}
}
\value{
A n by q matrix of simulated pixel counts where the sum of
  pixel counts in each row is identical.
}
\description{
The function is a wrapper to the cpp_boot function that
  eliminates the need for the user to generate random nunber seeds
  in c++. Note that the random seed generator get_seed() is an
  internal function in this package that is not available to
  the user but is described at:
  - https://www.r-bloggers.com/2018/09/using-rs-set-seed-to-set-seeds-for-use-in-c-c-including-rcpp/
}
