% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_boot.R
\name{resamp_wrapper}
\alias{resamp_wrapper}
\title{Parallel wrapper to bootstrap function.}
\usage{
resamp_wrapper(df, n, cores = 1)
}
\arguments{
\item{df}{A dataframe containing only pixel counts (no ID variable).}

\item{n}{The desired number of bootstrap simulations.}

\item{cores}{The desired number of cores. Anything greater than one
requires parallel processing capability using the doParallel package.
It is the users responsbility to know how many cores they have available
for the computations.}
}
\value{
A list of matrices, one matrix per row of df, with bootstrap
  pixel counts.
}
\description{
Parallel wrapper to bootstrap function.
}
