% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_counts}
\alias{soil_counts}
\title{Pixel counts by United State's County}
\format{
A data frame with 37,366 rows and 130 columns.

  \describe{
  \item{GEOID}{The unique county FIPS identifier.}
  \item{YEAR}{The year of crops for which classifications are made 2007-2019.}
  \item{Category_x}{The number of pixels in each county/year associated with
    crop type x.}
  }
}
\usage{
soil_counts
}
\description{
A dataset of pixel counts of land uses as provided by:
  https://www.nass.usda.gov/Research_and_Science/Cropland/docs/County_Pixel_Count.zip
}
\keyword{datasets}
